refactor(cache): refactor the cache
The current module design of the cache was beginning to get unruly.
Whlie rebasing `next` before 0.5.0, I noticed that a majority of the
merge conflicts were in the `lib` and `updates` files.

This PR moves accessors and update logic to files with a similar
resource, rather than having everything being in two files. It also
removes a few impleentations of `UpdateCache` that were empty or had no
functionality.
