+++ cache/in-memory/src/event/role.rs
@@ -1,0 +1,150 @@
+use crate::{config::ResourceType, InMemoryCache, UpdateCache};
+use std::collections::HashSet;
+use twilight_model::{
+    gateway::payload::{RoleCreate, RoleDelete, RoleUpdate},
+    guild::Role,
+    id::{GuildId, RoleId},
+};
+
+impl InMemoryCache {
+    /// Gets the set of roles in a guild.
+    ///
+    /// This is a O(m) operation, where m is the amount of roles in the guild.
+    /// This requires the [`GUILDS`] intent.
+    ///
+    /// [`GUILDS`]: ::twilight_model::gateway::Intents::GUILDS
+    pub fn guild_roles(&self, guild_id: GuildId) -> Option<HashSet<RoleId>> {
+        self.0.guild_roles.get(&guild_id).map(|r| r.clone())
+    }
+
+    /// Gets a role by ID.
+    ///
+    /// This is an O(1) operation. This requires the [`GUILDS`] intent.
+    ///
+    /// [`GUILDS`]: ::twilight_model::gateway::Intents::GUILDS
+    pub fn role(&self, role_id: RoleId) -> Option<Role> {
+        self.0.roles.get(&role_id).map(|r| r.data.clone())
+    }
+
+    pub(crate) fn cache_roles(&self, guild_id: GuildId, roles: impl IntoIterator<Item = Role>) {
+        for role in roles {
+            self.cache_role(guild_id, role);
+        }
+    }
+
+    fn cache_role(&self, guild_id: GuildId, role: Role) {
+        // Insert the role into the guild_roles map
+        self.0
+            .guild_roles
+            .entry(guild_id)
+            .or_default()
+            .insert(role.id);
+
+        // Insert the role into the all roles map
+        crate::upsert_guild_item(&self.0.roles, guild_id, role.id, role);
+    }
+
+    fn delete_role(&self, role_id: RoleId) {
+        if let Some((_, role)) = self.0.roles.remove(&role_id) {
+            if let Some(mut roles) = self.0.guild_roles.get_mut(&role.guild_id) {
+                roles.remove(&role_id);
+            }
+        }
+    }
+}
+
+impl UpdateCache for RoleCreate {
+    fn update(&self, cache: &InMemoryCache) {
+        if !cache.wants(ResourceType::ROLE) {
+            return;
+        }
+
+        crate::upsert_guild_item(
+            &cache.0.roles,
+            self.guild_id,
+            self.role.id,
+            self.role.clone(),
+        );
+    }
+}
+
+impl UpdateCache for RoleDelete {
+    fn update(&self, cache: &InMemoryCache) {
+        if !cache.wants(ResourceType::ROLE) {
+            return;
+        }
+
+        cache.delete_role(self.role_id);
+    }
+}
+
+impl UpdateCache for RoleUpdate {
+    fn update(&self, cache: &InMemoryCache) {
+        if !cache.wants(ResourceType::ROLE) {
+            return;
+        }
+
+        cache.cache_role(self.guild_id, self.role.clone());
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test;
+
+    #[test]
+    fn test_cache_role() {
+        let cache = InMemoryCache::new();
+
+        // Single inserts
+        {
+            // The role ids for the guild with id 1
+            let guild_1_role_ids = (1..=10).map(RoleId).collect::<Vec<_>>();
+            // Map the role ids to a test role
+            let guild_1_roles = guild_1_role_ids
+                .iter()
+                .copied()
+                .map(test::role)
+                .collect::<Vec<_>>();
+            // Cache all the roles using cache role
+            for role in guild_1_roles.clone() {
+                cache.cache_role(GuildId(1), role);
+            }
+
+            // Check for the cached guild role ids
+            let cached_roles = cache.guild_roles(GuildId(1)).unwrap();
+            assert_eq!(cached_roles.len(), guild_1_role_ids.len());
+            assert!(guild_1_role_ids.iter().all(|id| cached_roles.contains(id)));
+
+            // Check for the cached role
+            assert!(guild_1_roles
+                .into_iter()
+                .all(|role| cache.role(role.id).expect("Role missing from cache") == role))
+        }
+
+        // Bulk inserts
+        {
+            // The role ids for the guild with id 2
+            let guild_2_role_ids = (101..=110).map(RoleId).collect::<Vec<_>>();
+            // Map the role ids to a test role
+            let guild_2_roles = guild_2_role_ids
+                .iter()
+                .copied()
+                .map(test::role)
+                .collect::<Vec<_>>();
+            // Cache all the roles using cache roles
+            cache.cache_roles(GuildId(2), guild_2_roles.clone());
+
+            // Check for the cached guild role ids
+            let cached_roles = cache.guild_roles(GuildId(2)).unwrap();
+            assert_eq!(cached_roles.len(), guild_2_role_ids.len());
+            assert!(guild_2_role_ids.iter().all(|id| cached_roles.contains(id)));
+
+            // Check for the cached role
+            assert!(guild_2_roles
+                .into_iter()
+                .all(|role| cache.role(role.id).expect("Role missing from cache") == role))
+        }
+    }
+}
